+++ tests/fixt_jobmgr.py
@@ -1,10 +1,10 @@
 from copy import copy
 from time import sleep
 import pytest
-from pyppl import Proc as _Proc
+from pyppl import Proc as _Proc, Diot
+from pyppl.template import TemplateLiquid
 from pyppl.jobmgr import Jobmgr, STATES
 from pyppl.logger import logger
-from pyppl.utils import Box
 
 class Proc(dict):
 	OUT_VARTYPE    = _Proc.OUT_VARTYPE
@@ -28,10 +28,10 @@ class Proc(dict):
 		kwargs['errhow']  = 'terminate'
 		kwargs['errntry'] = 3
 		kwargs['forks']   = 1
+		kwargs['expect']  = TemplateLiquid('')
 		kwargs['size']    = 1
-		kwargs['config']  = Box(
-			_log = {}
-		)
+		kwargs['rc']      = [0]
+		kwargs['config']  = Diot(_log = {})
 		super(Proc, self).__init__(*args, **kwargs)
 
 	def __getattr__(self, item):